Don't overwrite an existing backup file

+++ setup.sh
@@ -32,8 +32,8 @@ fi
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 copy_with_backup () {
-	# If the destination file exists, back it up
-	if [ -f $2 ] then
+	# If the destination file exists, back it up--but only if the backup file does not already exist
+	if [ -f $2 ] && [ ! -f $2.rogueap-old ] then
 		mv -f ${2} ${2}.rogueap-old
 	cp -f ${1} ${2}
 }
